Tic Tac Toe for The Odin Project
This project allowed me to utilise functions and closures to create a Tic Tac Toe game.

There are minimal global variables in line with the project spec and the modular design means that each of the gameBoard, gameController and displayController are working independently of eachother and sharing with eachother.

gameBoard - controls the state of the board including the moving of pieces and resetting of the game
gameController - controls the flow of the game including storing the current players, the active player and declaring a winner
displayController - controls the display of the game based on the information in gameController and gameBoard.